♻️ refactor: rename cache keys constant module
