+++ apps/web/src/rpc/actions/challenging/UpdateChallengeVisibilityAction.ts
@@ -2,7 +2,7 @@ import type { ChallengingService } from '@stardust/core/challenging/interfaces'
 import type { Action, Call } from '@stardust/core/global/interfaces'
 import { Id } from '@stardust/core/global/structures'
 import { Challenge } from '@stardust/core/challenging/entities'
-import { CACHE_KEYS } from '@/constants/server-cache-keys'
+import { CACHE_KEYS } from '@/constants/cache-keys'
 
 type Request = {
   challengeId: string
