+++ apps/web/src/rpc/actions/challenging/tests/VoteChallengeAction.test.ts
@@ -3,7 +3,7 @@ import { mock, type Mock } from 'ts-jest-mocker'
 import type { ChallengingService } from '@stardust/core/challenging/interfaces'
 import type { Call } from '@stardust/core/global/interfaces'
 import { RestResponse } from '@stardust/core/global/responses'
-import { CACHE_KEYS } from '@/constants/server-cache-keys'
+import { CACHE_KEYS } from '@/constants/cache-keys'
 
 import { VoteChallengeAction } from '../VoteChallengeAction'
